+++ lib/services/app_installer_io.dart
@@ -6,6 +6,7 @@ import 'package:remote_storage/services/app_update_service.dart';
 import 'package:remote_storage/services/proxy_http_client.dart';
 import 'package:remote_storage/services/remote_storage_gateway.dart';
 import 'package:remote_storage/services/update_settings.dart';
+import 'package:remote_storage/models/remote_storage_config.dart';
 
 const bool kSupportsInAppInstall = true;
 
@@ -21,12 +22,15 @@ Future<String> downloadAndInstallAsset(
   void Function(int received, int total)? onProgress,
   UpdateNetworkConfig networkConfig = const UpdateNetworkConfig(),
   ProxyConfig proxyConfig = const ProxyConfig(),
+  RemoteStorageConfig? config,
 }) {
   return api.installApp(
     assetUrl: asset.downloadUrl,
     assetName: asset.name,
+    assetSize: asset.size,
     installerType: installerType,
     mirrorPrefix: networkConfig.mirrorPrefix,
+    config: config ?? RemoteStorageConfig.empty(),
     proxyMode: proxyConfig.mode,
     proxyType: proxyConfig.type,
     proxyHost: proxyConfig.host,