+++ lib/services/remote_storage_api_desktop_runtime.dart
@@ -12,8 +12,10 @@ mixin _RemoteStorageRuntimeApiMixin implements RemoteStorageGateway {
   Future<String> installApp({
     required String assetUrl,
     required String assetName,
+    required int assetSize,
     required String installerType,
     required String mirrorPrefix,
+    required RemoteStorageConfig config,
     required String proxyMode,
     required String proxyType,
     required String proxyHost,
@@ -24,8 +26,10 @@ mixin _RemoteStorageRuntimeApiMixin implements RemoteStorageGateway {
     final result = await runBridgeCall('install_app', <String, dynamic>{
       'assetUrl': assetUrl,
       'assetName': assetName,
+      'assetSize': assetSize,
       'installerType': installerType,
       'mirrorPrefix': mirrorPrefix,
+      'config': config.toJson(),
       'proxyMode': proxyMode,
       'proxyType': proxyType,
       'proxyHost': proxyHost,