+++ lib/services/remote_storage_api_web.dart
@@ -127,8 +127,10 @@ class RemoteStorageApi
   Future<String> installApp({
     required String assetUrl,
     required String assetName,
+    required int assetSize,
     required String installerType,
     required String mirrorPrefix,
+    required RemoteStorageConfig config,
     required String proxyMode,
     required String proxyType,
     required String proxyHost,