+++ test/widget_test.dart
@@ -170,8 +170,10 @@ class _FakeApi implements RemoteStorageGateway {
   Future<String> installApp({
     required String assetUrl,
     required String assetName,
+    required int assetSize,
     required String installerType,
     required String mirrorPrefix,
+    required RemoteStorageConfig config,
     required String proxyMode,
     required String proxyType,
     required String proxyHost,